requirements.txt
streamlit>=1.30
tensorflow>=2.13
numpy
pillow
